Make RouterOS more reliable with management IP config
Some RouterOS versions didn't detect interfaces quick enough to reliably accept an IP being assigned to ether1. The would reject adding the IP to an interface that didn't yet exist and so would be unreachable via management IP.
RouterOS 6.45.9 was particularly bad for this.

This change tells the router to wait for the interface to appear before trying to add the IP.