common/vrnetlab.py: added DHCP support for management interface
If CLAB_MGMT_DHCP environment variable is set, we assume that a DHCP client
inside of the VM will take care about setting the management IP and gateway.

From the implementation point of view the IPv4 and IPv6 addresses and gateways
are simply set to 'dhcp', and the platform specific `launch.py` script needs
to check it to decide if the interface needs to be configured manually or to
to enable dhcp (if not enabled by default).

The initial implementation has been tested with Broadcom Enterprise SONiC 4.4.0
using the "dell_sonic" kind (which actually is a rebranded Broadcom SONiC).
The primary focus is the ability to test the Zero Touch Provisioning feature of
SONiC but could also be used for testing PXE boot scenarios of other systems
using containerlab.

Note:
This feature is only working when CLAB_MGMT_PASSTHROUGH is true.

The feature needs to be defined explicitely within the node configuration
like this:
--------------------------------
env:
  CLAB_MGMT_PASSTHROUGH: "true"
  CLAB_MGMT_DHCP: "true"
--------------------------------